Do not allow to delete the last payment method of project
We don't allow to create virtual machines without billing info. But users
are able to add billing info, create vm, then delete payment method.

We need to have at least one payment method to charge user.

In the future, we may allow to delete the last payment method if project
doesn't have anything to charge.

+++ routes/web/project/billing.rb
@@ -95,6 +95,11 @@ class CloverWeb
         end
 
         r.delete true do
+          unless payment_method.billing_info.payment_methods.count > 1
+            response.status = 400
+            return {message: "You can't delete the last payment method of a project."}.to_json
+          end
+
           payment_method.destroy
 
           return {message: "Deleting #{payment_method.ubid}"}.to_json
